annotation/requirements_annotation.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0
opencv-python>=4.6.0
Pillow>=9.0.0
numpy>=1.21.0
gradio>=3.10.0  # Для визуального интерфейса
git+https://github.com/IDEA-Research/GroundingDINO.git